third_party/blink/web_tests/fast/dom/shadow/shadow-removechild-and-blur-event-expected.txt

CONSOLE WARNING: line 13: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Check focused shadow node dispatchs blur event on removeChild of host.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS logs.toString() is "start-removeChild,blur,end-removeChild"
PASS successfullyParsed is true

TEST COMPLETE
